python/requirements.txt
flask==3.0.0
pypdf==3.17.0
openai>=1.50.0
python-dotenv==1.0.0
